pyproject.toml
[tool.poetry]
name = "gcn-to-kafka"
version = "0.1.0"
description = "Pump GCN Notices to a Kafka broker"
authors = ["Leo Singer <leo.singer@ligo.org>"]
license = "NOSA"

[tool.poetry.dependencies]
click = "*"
python = "^3.6"
pygcn = "*"
kafka-python = "*"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
gcn-to-kafka = "gcn_to_kafka.cli:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
